3DC multi branches cleanup (#1885)
* In 3+DC, the mutable state may have multi-branches info. When a workflow is in cleanup, all branches need to be deleted. Currently, the logic is to remove all non-current branches (not max version), this is due to archival only archive current branch. In the future, when archival supports archive multiple branches. The cleanup can delete all branches.

+++ service/history/timerQueueProcessorBase.go
@@ -617,7 +617,7 @@ func (t *timerQueueProcessorBase) deleteWorkflow(task *persistence.TimerTaskInfo
 		return err
 	}
 
-	if err := t.deleteWorkflowHistory(task, msBuilder); err != nil {
+	if err := t.deleteWorkflowHistory(task, msBuilder, true); err != nil {
 		return err
 	}
 
@@ -666,6 +666,9 @@ func (t *timerQueueProcessorBase) archiveWorkflow(task *persistence.TimerTaskInf
 	if err := t.deleteWorkflowExecution(task); err != nil {
 		return err
 	}
+	if err := t.deleteWorkflowHistory(task, msBuilder, false); err != nil {
+		return err
+	}
 	if err := t.deleteWorkflowVisibility(task); err != nil {
 		return err
 	}
@@ -697,7 +700,7 @@ func (t *timerQueueProcessorBase) deleteCurrentWorkflowExecution(task *persisten
 	return backoff.Retry(op, persistenceOperationRetryPolicy, common.IsPersistenceTransientError)
 }
 
-func (t *timerQueueProcessorBase) deleteWorkflowHistory(task *persistence.TimerTaskInfo, msBuilder mutableState) error {
+func (t *timerQueueProcessorBase) deleteWorkflowHistory(task *persistence.TimerTaskInfo, msBuilder mutableState, shouldDeleteCurrent bool) error {
 	domainID, workflowExecution := t.getDomainIDAndWorkflowExecution(task)
 	op := func() error {
 		if msBuilder.GetEventStoreVersion() == persistence.EventStoreVersionV2 {
@@ -708,8 +711,30 @@ func (t *timerQueueProcessorBase) deleteWorkflowHistory(task *persistence.TimerT
 				tag.TaskID(task.GetTaskID()),
 				tag.FailoverVersion(task.GetVersion()),
 				tag.TaskType(task.GetTaskType()))
-			return persistence.DeleteWorkflowExecutionHistoryV2(t.historyService.historyV2Mgr, msBuilder.GetCurrentBranch(), common.IntPtr(t.shard.GetShardID()), logger)
+			versionHistories := msBuilder.GetAllVersionHistories()
+
+			if versionHistories != nil {
+			CleanupLoop:
+				for idx, versionHistory := range versionHistories.Histories {
+					if !shouldDeleteCurrent && int32(idx) == versionHistories.CurrentBranch {
+						continue CleanupLoop
+					}
+					if err := persistence.DeleteWorkflowExecutionHistoryV2(
+						t.historyService.historyV2Mgr,
+						versionHistory.BranchToken,
+						common.IntPtr(t.shard.GetShardID()),
+						logger); err != nil {
+						return err
+					}
+					msBuilder.DeleteVersionHistory(idx)
+					//TODO: update mutable state with deleted version histories
+				}
+			}
 		}
+		if !shouldDeleteCurrent {
+			return nil
+		}
+
 		return t.historyService.historyMgr.DeleteWorkflowExecutionHistory(
 			&persistence.DeleteWorkflowExecutionHistoryRequest{
 				DomainID:  domainID,
